Home Fragment code cleanup
