+++ app/src/main/java/com/example/devexercise/util/CountryClick.kt
@@ -1,0 +1,7 @@
+package com.example.devexercise.util
+
+import com.example.devexercise.repository.CountryModel
+
+class CountryClick(val clickListener: (CountryModel) -> Unit){
+    fun onClick(country: CountryModel) = clickListener(country)
+}
